+++ src-api/app/studium/core/datacontroller.py
@@ -1,0 +1,4 @@
+from typing import Optional
+from cloudant import CouchDB
+
+# Todo: Reimplement